update the ui
