+++ feature/trip-planner/ui/src/commonMain/kotlin/xyz/ksharma/krail/trip/planner/ui/departureboard/DepartureBoardStopSection.kt
@@ -8,23 +8,15 @@ import androidx.compose.animation.core.tween
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.lazy.LazyListScope
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.saveable.rememberSaveable
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.rotate
@@ -33,24 +25,19 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import kotlinx.collections.immutable.ImmutableList
 import kotlinx.collections.immutable.persistentListOf
-import kotlinx.collections.immutable.toImmutableList
 import krail.feature.trip_planner.ui.generated.resources.Res
 import krail.feature.trip_planner.ui.generated.resources.ic_arrow_down
 import org.jetbrains.compose.resources.painterResource
 import xyz.ksharma.krail.departures.ui.state.DeparturesState
 import xyz.ksharma.krail.departures.ui.state.model.StopDeparture
-import xyz.ksharma.krail.taj.components.Divider
 import xyz.ksharma.krail.taj.components.Text
-import xyz.ksharma.krail.taj.components.TextButton
 import xyz.ksharma.krail.taj.modifier.klickable
 import xyz.ksharma.krail.taj.preview.PreviewComponent
 import xyz.ksharma.krail.taj.theme.KrailTheme
 import xyz.ksharma.krail.taj.theme.KrailThemeStyle
 import xyz.ksharma.krail.taj.theme.PreviewTheme
 import xyz.ksharma.krail.taj.themeBackgroundColor
-import xyz.ksharma.krail.trip.planner.ui.components.DepartureRowList
-import xyz.ksharma.krail.trip.planner.ui.components.DeparturesErrorContent
-import xyz.ksharma.krail.trip.planner.ui.components.LinesServedRow
+import xyz.ksharma.krail.trip.planner.ui.components.DepartureBoardBody
 import xyz.ksharma.krail.trip.planner.ui.components.loading.AnimatedDots
 import xyz.ksharma.krail.trip.planner.ui.savedtrips.StopDepartureBoardEntry
 
@@ -61,9 +48,9 @@ import xyz.ksharma.krail.trip.planner.ui.savedtrips.StopDepartureBoardEntry
  *  - `"${stopId}_header"` — sticky tappable card
  *  - `"${stopId}_content"` — collapsible content with services row + departure list
  *
- * The content item hosts [DepartureBoardAccordionContent] which owns the per-stop filter state.
- * Filter state naturally resets when the item collapses because [remember] is keyed to the
- * item's composition lifetime.
+ * The content item hosts [DepartureBoardAccordionContent] which delegates all state
+ * to [DepartureBoardBody]. Filter and previous-departures state resets automatically
+ * when the item leaves composition (section collapses).
  */
 fun LazyListScope.departureBoardAccordionSection(
     entry: StopDepartureBoardEntry,
@@ -168,11 +155,9 @@ internal fun DepartureBoardAccordionSectionHeader(
 /**
  * Content area for an expanded stop section.
  *
- * Owns the per-stop line filter state locally. Filter resets automatically when this
- * composable leaves composition (i.e. when the section collapses), so re-expanding
- * always starts with all services shown.
+ * All filter and previous-departures state is owned by [DepartureBoardBody] and naturally
+ * resets when this composable leaves composition (i.e. when the section collapses).
  */
-@Suppress("CyclomaticComplexMethod") // todo split this method
 @Composable
 private fun DepartureBoardAccordionContent(
     stopId: String,
@@ -181,99 +166,12 @@ private fun DepartureBoardAccordionContent(
     onRefreshStop: (String) -> Unit,
     modifier: Modifier = Modifier,
 ) {
-    // Empty string = no filter. Non-null String is guaranteed safe through rememberSaveable
-    // (Bundle / SavedStateHandle only reliably round-trip primitives and String).
-    var selectedLineKey by rememberSaveable { mutableStateOf("") }
-    val selectedLine: String? = selectedLineKey.ifEmpty { null }
-
-    var showPrevious by remember { mutableStateOf(false) }
-    // Tracks whether a previous-departures fetch has been requested but isPreviousLoading
-    // hasn't arrived yet. Without this, the "No previous departures" text flashes for one
-    // frame before the repository sets isPreviousLoading = true.
-    var previousRequested by remember { mutableStateOf(false) }
-
-    // Clear the local guard once the repository acknowledges the request (loading starts/ends).
-    LaunchedEffect(state.isPreviousLoading) {
-        if (!state.isPreviousLoading) previousRequested = false
-    }
-
-    val filteredDepartures = remember(state.departures, selectedLine) {
-        if (selectedLine == null) {
-            state.departures
-        } else {
-            state.departures.filter { it.lineNumber == selectedLine }.toImmutableList()
-        }
-    }
-
-    val filteredPreviousDepartures = remember(state.previousDepartures, selectedLine) {
-        if (selectedLine == null) {
-            state.previousDepartures
-        } else {
-            state.previousDepartures.filter { it.lineNumber == selectedLine }.toImmutableList()
-        }
-    }
-
-    Column(modifier = modifier) {
-        when {
-            state.isLoading -> SectionLoadingContent()
-            state.isError -> DeparturesErrorContent(onRetry = { onRefreshStop(stopId) })
-            state.departures.isEmpty() -> SectionEmptyContent(hasActiveFilter = false)
-            else -> {
-                LinesServedRow(
-                    departures = state.departures,
-                    selectedLine = selectedLine,
-                    onLineSelect = { selectedLineKey = it ?: "" },
-                )
-
-                // "Show previous / Hide previous" toggle
-                ShowPreviousButton(
-                    modifier = Modifier.padding(top = 4.dp, bottom = 12.dp),
-                    showPrevious = showPrevious,
-                    onClick = {
-                        showPrevious = !showPrevious
-                        if (showPrevious && state.previousDepartures.isEmpty() && !state.isPreviousLoading) {
-                            previousRequested = true
-                            onLoadPreviousDepartures(stopId)
-                        }
-                    },
-                )
-
-                when {
-                    // Previous fetch in progress (or requested but not yet acknowledged) —
-                    // show inline loader above upcoming departures.
-                    showPrevious && (state.isPreviousLoading || previousRequested) -> {
-                        SectionLoadingContent()
-                        when {
-                            filteredDepartures.isEmpty() -> SectionEmptyContent(hasActiveFilter = true)
-                            else -> DepartureRowList(departures = filteredDepartures)
-                        }
-                    }
-                    // Previous fetched but none match current filter — show message then upcoming
-                    showPrevious && filteredPreviousDepartures.isEmpty() -> {
-                        Text(
-                            text = "No previous departures in the last ${state.previousWindowMinutes} minutes.",
-                            style = KrailTheme.typography.bodyMedium,
-                            color = KrailTheme.colors.softLabel,
-                            modifier = Modifier.padding(horizontal = 16.dp, vertical = 12.dp),
-                        )
-                        when {
-                            filteredDepartures.isEmpty() -> SectionEmptyContent(hasActiveFilter = true)
-                            else -> DepartureRowList(departures = filteredDepartures)
-                        }
-                    }
-                    // Previous + upcoming combined — date labels flow as one unified list
-                    showPrevious -> DepartureRowList(
-                        departures = remember(filteredPreviousDepartures, filteredDepartures) {
-                            (filteredPreviousDepartures + filteredDepartures).toImmutableList()
-                        },
-                    )
-                    filteredDepartures.isEmpty() -> SectionEmptyContent(hasActiveFilter = true)
-                    else -> DepartureRowList(departures = filteredDepartures)
-                }
-            }
-        }
-        Spacer(modifier = Modifier.height(8.dp))
-    }
+    DepartureBoardBody(
+        state = state,
+        onRetry = { onRefreshStop(stopId) },
+        onLoadPreviousDepartures = { onLoadPreviousDepartures(stopId) },
+        modifier = modifier,
+    )
 }
 
 // ── Standalone composable (previews only) ────────────────────────────────────
@@ -312,54 +210,6 @@ fun DepartureBoardAccordionSection(
     }
 }
 
-// ── Show previous button ──────────────────────────────────────────────────────
-
-@Composable
-private fun ShowPreviousButton(
-    showPrevious: Boolean,
-    onClick: () -> Unit,
-    modifier: Modifier = Modifier,
-) {
-    Box(
-        modifier = modifier.fillMaxWidth(),
-        contentAlignment = Alignment.Center,
-    ) {
-        TextButton(onClick = onClick) {
-            Text(text = if (showPrevious) "Hide previous departures" else "Show previous departures")
-        }
-    }
-}
-
-// ── Private content slots ─────────────────────────────────────────────────────
-
-@Composable
-private fun SectionLoadingContent() {
-    Box(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(vertical = 24.dp),
-        contentAlignment = Alignment.Center,
-    ) {
-        AnimatedDots(
-            modifier = Modifier.size(width = 64.dp, height = 24.dp),
-            color = KrailTheme.colors.onSurface,
-        )
-    }
-}
-
-@Composable
-private fun SectionEmptyContent(hasActiveFilter: Boolean) {
-    Text(
-        text = if (hasActiveFilter) {
-            "No departures match the selected line."
-        } else {
-            "No upcoming departures found."
-        },
-        style = KrailTheme.typography.bodyMedium,
-        color = KrailTheme.colors.softLabel,
-        modifier = Modifier.padding(horizontal = 16.dp, vertical = 16.dp),
-    )
-}
 
 // ── Previews ──────────────────────────────────────────────────────────────────
