+++ feature/trip-planner/ui/src/commonMain/kotlin/xyz/ksharma/krail/trip/planner/ui/searchstop/map/StopDetailsBottomSheet.kt
@@ -55,7 +55,6 @@ fun StopDetailsBottomSheet(
                 stopId = stop.stopId,
                 state = departuresState,
                 onEvent = departuresViewModel::onEvent,
-                modifier = Modifier.padding(horizontal = 16.dp),
             )
 
             Spacer(modifier = Modifier.height(16.dp))
